consume_row_start: change type of clustering key
Clustering key in 3.x format is stored differently
so it's easier to create a vector of temporary buffers
instead of a single block of concatenated bytes.

Each temporary buffer stores a value of a single
clustering column.

This is because the way clustering key is stored on disk
in SSTables 3.x is not the same as the way we store it
internally.

This means that we have to first read a value of every
clustering column into temporary_buffer and only then
we can create clustering key using a vector of those
temporary buffers.

Signed-off-by: Piotr Jastrzebski <piotr@scylladb.com>