sstables: Fix calculation of previous row size for SSTables 3.x
The previous code incorrectly calculated sizes of previous rows while
writing SSTables in 3.x ('m') format.
The problem with detecting this issue was that neither sstabledump nor
Cassandra 3.x itself use those values, as of today, they are simply
ignored when data is read from files.

Still, we want to be compatible and write correct values as they may be
of use in the future.

Signed-off-by: Vladimir Krivopalov <vladimir@scylladb.com>